New trait WP_Rocket\Busting\AdminNotices
