+++ inc/classes/subscriber/class-facebook-tracking-cache-busting-subscriber.php
@@ -3,6 +3,7 @@
 namespace WP_Rocket\Subscriber;
 
 use WP_Rocket\Event_Management\Subscriber_Interface;
+use WP_Rocket\Busting\AdminNotices;
 use WP_Rocket\Busting\Busting_Factory;
 use WP_Rocket\Admin\Options_Data as Options;
 
@@ -13,6 +14,7 @@ use WP_Rocket\Admin\Options_Data as Options;
  * @author Grégory Viguier
  */
 class Facebook_Tracking_Cache_Busting_Subscriber implements Subscriber_Interface {
+	use AdminNotices;
 
 	/**
 	 * Name of the cron.
@@ -22,6 +24,7 @@ class Facebook_Tracking_Cache_Busting_Subscriber implements Subscriber_Interface
 	 * @author Grégory Viguier
 	 */
 	const CRON_NAME = 'rocket_facebook_tracking_cache_update';
+
 	/**
 	 * Instance of the Busting Factory class.
 	 *
@@ -41,6 +44,20 @@ class Facebook_Tracking_Cache_Busting_Subscriber implements Subscriber_Interface
 	 * @author Grégory Viguier
 	 */
 	private $options;
+
+	/**
+	 * Busting types.
+	 *
+	 * @var string
+	 */
+	private $busting_types = [ 'fbsdk', 'fbpix' ];
+
+	/**
+	 * Vendor name.
+	 *
+	 * @var string
+	 */
+	private $vendor_name = 'Facebook';
 
 	/**
 	 * Constructor.
@@ -182,63 +199,6 @@ class Facebook_Tracking_Cache_Busting_Subscriber implements Subscriber_Interface
 		$html = $this->busting_factory->type( 'fbsdk' )->replace_url( $html );
 
 		return $this->busting_factory->type( 'fbpix' )->replace_url( $html );
-	}
-
-	/**
-	 * Display an admin notice if the cache folder is not writable.
-	 *
-	 * @since  3.6
-	 * @author Grégory Viguier
-	 */
-	public function busting_dir_not_writable_admin_notice() {
-		if ( ! $this->is_busting_active() || ! current_user_can( 'rocket_manage_options' ) ) {
-			return;
-		}
-
-		$dir_paths = [
-			$this->busting_factory->type( 'fbsdk' )->get_busting_dir_path(),
-			$this->busting_factory->type( 'fbpix' )->get_busting_dir_path(),
-		];
-
-		$dir_paths  = array_unique( $dir_paths );
-		$filesystem = rocket_direct_filesystem();
-
-		foreach ( $dir_paths as $i => $dir_path ) {
-			if ( ! $filesystem->exists( $dir_path ) ) {
-				rocket_mkdir_p( $dir_path );
-			}
-			if ( $filesystem->exists( $dir_path ) && $filesystem->is_writable( $dir_path ) ) {
-				unset( $dir_paths[ $i ] );
-			} else {
-				$dir_paths[ $i ] = '<code>' . esc_html( trim( str_replace( ABSPATH, '', $dir_path ), '/' ) ) . '</code>';
-			}
-		}
-
-		if ( ! $dir_paths ) {
-			return;
-		}
-
-		$message  = '<strong>' . __( 'WP Rocket: ', 'rocket' ) . '</strong>';
-		$message .= sprintf(
-			/* translators: %s is a list of folder paths. */
-			_n( 'The folder %s used to cache Facebook tracking scripts could not be created or is missing writing permissions.', 'The folders %s used to cache Facebook tracking scripts could not be created or are missing writing permissions.', count( $dir_paths ), 'rocket' ),
-			wp_sprintf_l( '%l', $dir_paths )
-		);
-		$message .= '<br>' . sprintf(
-			/* translators: This is a doc title! %1$s = opening link; %2$s = closing link */
-			__( 'Troubleshoot: %1$sHow to make system files writeable%2$s', 'rocket' ),
-			/* translators: Documentation exists in EN, DE, FR, ES, IT; use loaclised URL if applicable */
-			'<a href="' . __( 'https://docs.wp-rocket.me/article/626-how-to-make-system-files-htaccess-wp-config-writeable/?utm_source=wp_plugin&utm_medium=wp_rocket', 'rocket' ) . '" target="_blank">',
-			'</a>'
-		);
-
-		rocket_notice_html(
-			[
-				'status'      => 'error',
-				'dismissible' => '',
-				'message'     => $message,
-			]
-		);
 	}
 
 	/**
